Move london underground coordinator to its own file
